ajuste interfaces de compras
